Move log-ingest domain entities into core/entities
The AttemptStream and JobAccounting domain entities (and the StreamState /
StreamCloseReason value types) were defined alongside their Drizzle tables in
db/schema, mixing the persistence shape with the domain shape. Move the entity
definitions into core/entities; the schema files keep the tables, inferred DB
row types, and the toX() row-to-domain mappers, which now import the entities
from core. The core layer only ever sees domain entities, never DB types.

Refs ENG-439
